# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusExchanges = {[key: string]: string};

export interface HealthStatus {
  status: string;
  exchanges?: HealthStatusExchanges;
  active_subscriptions?: number;
}

export interface SymbolsResponse {
  symbols: string[];
  intervals: string[];
  default_intervals?: string[];
}

export interface Candle {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
  /** @nullable */
  is_closed?: boolean | null;
}

export interface CandlesResponse {
  symbol: string;
  interval: string;
  candles: Candle[];
}

export interface RsiPoint {
  time: number;
  /** @nullable */
  value?: number | null;
}

export interface RsiResponse {
  symbol: string;
  interval: string;
  period: number;
  source: string;
  rsi: RsiPoint[];
}

export interface BollingerBandsPoint {
  time: number;
  /** @nullable */
  upper?: number | null;
  /** @nullable */
  middle?: number | null;
  /** @nullable */
  lower?: number | null;
}

export interface StochRsiPoint {
  time: number;
  /** @nullable */
  k?: number | null;
  /** @nullable */
  d?: number | null;
}

export type DivergenceMarkerType = typeof DivergenceMarkerType[keyof typeof DivergenceMarkerType];


export const DivergenceMarkerType = {
  bullish: 'bullish',
  bearish: 'bearish',
} as const;

export interface DivergenceMarker {
  time: number;
  type: DivergenceMarkerType;
  price?: number;
  rsi_value?: number;
}

export interface MultiTimeframeRsiEntry {
  interval: string;
  /** @nullable */
  value?: number | null;
}

export interface RsiAdvancedResponse {
  symbol: string;
  interval: string;
  period: number;
  source: string;
  rsi: RsiPoint[];
  sma_rsi?: RsiPoint[];
  ema_rsi?: RsiPoint[];
  wma_rsi?: RsiPoint[];
  bollinger_bands?: BollingerBandsPoint[];
  stoch_rsi?: StochRsiPoint[];
  divergences?: DivergenceMarker[];
  mtf_rsi?: MultiTimeframeRsiEntry[];
}

export interface ErrorResponse {
  error: string;
}

export type GetCandlesParams = {
/**
 * Normalized symbol, e.g. BTC/USD
 */
symbol: string;
/**
 * Candle interval, e.g. 1m, 5m, 1h
 */
interval: string;
/**
 * Number of candles to return
 */
limit?: number;
/**
 * Fetch candles before this unix timestamp (exclusive)
 */
before?: number;
};

export type GetRsiParams = {
symbol: string;
interval: string;
period?: number;
source?: GetRsiSource;
limit?: number;
/**
 * Fetch RSI candles before this unix timestamp (exclusive)
 */
before?: number;
};

export type GetRsiSource = typeof GetRsiSource[keyof typeof GetRsiSource];


export const GetRsiSource = {
  close: 'close',
  open: 'open',
  high: 'high',
  low: 'low',
  hl2: 'hl2',
  hlc3: 'hlc3',
  ohlc4: 'ohlc4',
} as const;

export type GetRsiAdvancedParams = {
symbol: string;
interval: string;
period?: number;
source?: string;
limit?: number;
include_sma?: boolean;
sma_period?: number;
include_ema?: boolean;
ema_period?: number;
include_wma?: boolean;
wma_period?: number;
include_stoch_rsi?: boolean;
include_bb?: boolean;
include_divergence?: boolean;
include_mtf?: boolean;
/**
 * Fetch RSI candles before this unix timestamp (exclusive)
 */
before?: number;
};