# lib/api-zod/src/generated/types/bollingerBandsPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */

export interface BollingerBandsPoint {
  time: number;
  /** @nullable */
  upper?: number | null;
  /** @nullable */
  middle?: number | null;
  /** @nullable */
  lower?: number | null;
}