# lib/api-zod/src/generated/types/candle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */

export interface Candle {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
  /** @nullable */
  is_closed?: boolean | null;
}