# lib/api-zod/src/generated/types/candlesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';

export interface CandlesResponse {
  symbol: string;
  interval: string;
  candles: Candle[];
}