# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusExchanges } from './healthStatusExchanges';

export interface HealthStatus {
  status: string;
  exchanges?: HealthStatusExchanges;
  active_subscriptions?: number;
}