# lib/api-zod/src/generated/types/rsiAdvancedResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */
import type { BollingerBandsPoint } from './bollingerBandsPoint';
import type { DivergenceMarker } from './divergenceMarker';
import type { MultiTimeframeRsiEntry } from './multiTimeframeRsiEntry';
import type { RsiPoint } from './rsiPoint';
import type { StochRsiPoint } from './stochRsiPoint';

export interface RsiAdvancedResponse {
  symbol: string;
  interval: string;
  period: number;
  source: string;
  rsi: RsiPoint[];
  sma_rsi?: RsiPoint[];
  ema_rsi?: RsiPoint[];
  wma_rsi?: RsiPoint[];
  bollinger_bands?: BollingerBandsPoint[];
  stoch_rsi?: StochRsiPoint[];
  divergences?: DivergenceMarker[];
  mtf_rsi?: MultiTimeframeRsiEntry[];
}