# lib/api-zod/src/generated/types/rsiResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */
import type { RsiPoint } from './rsiPoint';

export interface RsiResponse {
  symbol: string;
  interval: string;
  period: number;
  source: string;
  rsi: RsiPoint[];
}