# lib/api-zod/src/generated/types/stochRsiPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */

export interface StochRsiPoint {
  time: number;
  /** @nullable */
  k?: number | null;
  /** @nullable */
  d?: number | null;
}