# lib/api-zod/src/generated/types/symbolsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading app API — candles, symbols, indicators
 * OpenAPI spec version: 0.1.0
 */

export interface SymbolsResponse {
  symbols: string[];
  intervals: string[];
  default_intervals?: string[];
}